chore: Rename signUpPage.tsx to signup.tsx and update imports in page.tsx
